Add new Cesium3DTilesContent library.
