+++ CMakeLists.txt
@@ -148,6 +148,7 @@ add_subdirectory(CesiumAsync)
 add_subdirectory(Cesium3DTiles)
 add_subdirectory(Cesium3DTilesReader)
 add_subdirectory(Cesium3DTilesWriter)
+add_subdirectory(Cesium3DTilesContent)
 add_subdirectory(Cesium3DTilesSelection)
 add_subdirectory(CesiumIonClient)
