+++ Cesium3DTilesContent/include/Cesium3DTilesContent/BinaryToGltfConverter.h
@@ -1,17 +1,20 @@
 #pragma once
 
-#include <Cesium3DTilesSelection/GltfConverterResult.h>
-#include <CesiumGltf/Model.h>
+#include <Cesium3DTilesContent/GltfConverterResult.h>
 #include <CesiumGltfReader/GltfReader.h>
 
 #include <gsl/span>
 
-#include <optional>
+#include <cstddef>
 
 namespace Cesium3DTilesSelection {
-struct PntsToGltfConverter {
+struct BinaryToGltfConverter {
+public:
   static GltfConverterResult convert(
-      const gsl::span<const std::byte>& pntsBinary,
+      const gsl::span<const std::byte>& gltfBinary,
       const CesiumGltfReader::GltfReaderOptions& options);
+
+private:
+  static CesiumGltfReader::GltfReader _gltfReader;
 };
 } // namespace Cesium3DTilesSelection
