+++ Cesium3DTilesContent/include/Cesium3DTilesContent/PntsToGltfConverter.h
@@ -1,6 +1,7 @@
 #pragma once
 
-#include <Cesium3DTilesSelection/GltfConverterResult.h>
+#include "GltfConverterResult.h"
+
 #include <CesiumGltf/Model.h>
 #include <CesiumGltfReader/GltfReader.h>
 
@@ -9,9 +10,9 @@
 #include <optional>
 
 namespace Cesium3DTilesSelection {
-struct B3dmToGltfConverter {
+struct PntsToGltfConverter {
   static GltfConverterResult convert(
-      const gsl::span<const std::byte>& b3dmBinary,
+      const gsl::span<const std::byte>& pntsBinary,
       const CesiumGltfReader::GltfReaderOptions& options);
 };
 } // namespace Cesium3DTilesSelection
