+++ Cesium3DTilesContent/src/BatchTableHierarchyPropertyValues.cpp
