+++ Cesium3DTilesContent/src/BatchTableHierarchyPropertyValues.h
