+++ Cesium3DTilesContent/test/TestUpgradeBatchTableToExtStructuralMetadata.cpp
