+++ Cesium3DTilesSelection/CMakeLists.txt
@@ -44,6 +44,7 @@ target_include_directories(
 target_link_libraries(Cesium3DTilesSelection
     PUBLIC
         Cesium3DTiles
+        Cesium3DTilesContent
         Cesium3DTilesReader
         CesiumAsync
         CesiumGeospatial
